перенёс данные и функции в отдельные файлы
