+++ data.php
@@ -1,0 +1,41 @@
+<?php
+    $categories = ['Входящие', 'Учёба', 'Работа', 'Домашние дела', 'Авто'];
+    $task_list = [
+        [
+            'task' => 'Собеседование в IT компании',
+            'date' => '01.12.2019',
+            'category' => 'Работа',
+            'status' => false
+        ],
+        [
+            'task' => 'Выполнить тестовое задание',
+            'date' => '25.12.2019',
+            'category' => 'Работа',
+            'status' => false
+        ],
+        [
+            'task' => 'Сделать задание первого раздела',
+            'date' => '21.12.2019',
+            'category' => 'Учеба',
+            'status' => true
+        ],
+        [
+            'task' => 'Встреча с другом',
+            'date' => '22.12.2019',
+            'category' => 'Входящие',
+            'status' => false
+        ],
+        [
+            'task' => 'Купить корм для кота',
+            'date' => null,
+            'category' => 'Домашние дела',
+            'status' => false
+        ],
+        [
+            'task' => 'Заказать пиццу',
+            'date' => null,
+            'category' => 'Домашние дела',
+            'status' => false
+        ]
+    ];
+?>
